# openapi.yml
openapi: 3.0.3
info:
  title: PrimeraApiCasa - Biblioteca API
  description: |
    REST API para gestión de una biblioteca. Usa autenticación JWT.

    **Roles:**
    - `USER` — puede consultar el catálogo (autores, libros, detalles, categorías, libro-categorías)
    - `ADMIN` — acceso completo, incluyendo escritura en el catálogo y gestión de roles y usuarios

    **Autenticación:** Los endpoints protegidos requieren el header `Authorization: Bearer <token>`.
    El token se obtiene en `/auth/login` o `/auth/register`.
  version: 1.0.0
servers:
  - url: http://localhost:8080/micro
    description: Servidor local
security:
  - bearerAuth: []
...
paths:
...
  /libros:
    get:
      tags: [Libros]
      summary: Listar todos los libros
      description: Requiere rol `USER` o `ADMIN`.
      responses:
        '200':
          description: Lista de libros
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LibroResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Libros]
      summary: Crear un libro
      description: Requiere rol `ADMIN`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibroCreateRequest'
            example:
              titulo: Cien años de soledad
              anioPublicacion: 1967
              autorId: 1
      responses:
        '201':
          description: Libro creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibroResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /libros/{id}:
    get:
      tags: [Libros]
      summary: Obtener libro por ID
      description: Requiere rol `USER` o `ADMIN`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Libro encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibroResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    put:
      tags: [Libros]
      summary: Actualizar un libro
      description: Requiere rol `ADMIN`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibroCreateRequest'
            example:
              titulo: Cien años de soledad
              anioPublicacion: 1967
              autorId: 1
      responses:
        '200':
          description: Libro actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibroResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      tags: [Libros]
      summary: Eliminar un libro
      description: Requiere rol `ADMIN`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Libro eliminado
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /libros/autor/{autorId}:
    get:
      tags: [Libros]
      summary: Listar libros por autor
      description: Requiere rol `USER` o `ADMIN`.
      parameters:
        - name: autorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del autor
      responses:
        '200':
          description: Lista de libros del autor
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LibroResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ─── DETALLES ────────────────────────────────────────────────────────────────
  /detalles/{id}:
    get:
      tags: [Detalles]
      summary: Obtener detalle de libro por ID
      description: Requiere rol `USER` o `ADMIN`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Detalle encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetalleLibroResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    put:
      tags: [Detalles]
      summary: Actualizar detalle de libro
      description: Requiere rol `ADMIN`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetalleLibroUpdateRequest'
            example:
              isbn: 978-3-16-148410-0
              numPaginas: 432
              idioma: Español
      responses:
        '200':
          description: Detalle actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetalleLibroResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      tags: [Detalles]
      summary: Eliminar detalle de libro
      description: Requiere rol `ADMIN`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Detalle eliminado
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /detalles/libro/{libroId}:
    get:
      tags: [Detalles]
      summary: Obtener detalle por ID de libro
      description: Requiere rol `USER` o `ADMIN`.
      parameters:
        - name: libroId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del libro
      responses:
        '200':
          description: Detalle encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetalleLibroResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /detalles:
    post:
      tags: [Detalles]
      summary: Crear detalle de libro
      description: Requiere rol `ADMIN`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DetalleLibroCreateRequest'
            example:
              isbn: 978-3-16-148410-0
              numPaginas: 432
              idioma: Español
              libroId: 1
      responses:
        '201':
          description: Detalle creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetalleLibroResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ─── CATEGORIAS ──────────────────────────────────────────────────────────────
...
  /libro-categorias/libro/{libroId}:
    get:
      tags: [LibroCategorias]
      summary: Obtener categorías de un libro
      description: Requiere rol `USER` o `ADMIN`.
      parameters:
        - name: libroId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del libro
      responses:
        '200':
          description: Lista de asociaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LibroCategoriaResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /libro-categorias/categoria/{categoriaId}:
    get:
      tags: [LibroCategorias]
      summary: Obtener libros de una categoría
      description: Requiere rol `USER` o `ADMIN`.
      parameters:
        - name: categoriaId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID de la categoría
      responses:
        '200':
          description: Lista de asociaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LibroCategoriaResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /libro-categorias:
    post:
      tags: [LibroCategorias]
      summary: Asociar libro con categoría
      description: Requiere rol `ADMIN`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibroCategoriaCreateRequest'
            example:
              libroId: 1
              categoriaId: 2
              prioridad: 1
              comentario: Categoría principal
      responses:
        '201':
          description: Asociación creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibroCategoriaResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /libro-categorias/{libroId}/{categoriaId}:
    delete:
      tags: [LibroCategorias]
      summary: Eliminar asociación libro-categoría
      description: Requiere rol `ADMIN`.
      parameters:
        - name: libroId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del libro
        - name: categoriaId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID de la categoría
      responses:
        '204':
          description: Asociación eliminada
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ─── ROLES ───────────────────────────────────────────────────────────────────
...
  /usuario-roles/rol/{rolId}:
    get:
      tags: [UsuarioRoles]
      summary: Obtener usuarios con un rol
      description: Requiere rol `ADMIN`.
      parameters:
        - name: rolId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del rol
      responses:
        '200':
          description: Lista de asignaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsuarioRolResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: ID del recurso
  responses:
    Unauthorized:
      description: Token ausente o inválido
    NotFound:
      description: Recurso no encontrado
    Forbidden:
      description: No tienes permisos para este recurso
  schemas:
...
    LibroCreateRequest:
      type: object
      required: [titulo, autorId]
      properties:
        titulo:
          type: string
          maxLength: 200
          example: Cien años de soledad
        anioPublicacion:
          type: integer
          example: 1967
        autorId:
          type: integer
          format: int64
          example: 1
    LibroResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        titulo:
          type: string
          example: Cien años de soledad
        anioPublicacion:
          type: integer
          example: 1967
        autorId:
          type: integer
          format: int64
          example: 1
        autorNombre:
          type: string
          example: Gabriel García Márquez
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Detalles
    DetalleLibroCreateRequest:
      type: object
      required: [isbn, libroId]
      properties:
        isbn:
          type: string
          maxLength: 20
          example: 978-3-16-148410-0
        numPaginas:
          type: integer
          example: 432
        idioma:
          type: string
          maxLength: 50
          example: Español
        libroId:
          type: integer
          format: int64
          example: 1
    DetalleLibroUpdateRequest:
      type: object
      properties:
        isbn:
          type: string
          maxLength: 20
          example: 978-3-16-148410-0
        numPaginas:
          type: integer
          example: 432
        idioma:
          type: string
          maxLength: 50
          example: Español
    DetalleLibroResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        isbn:
          type: string
          example: 978-3-16-148410-0
        numPaginas:
          type: integer
          example: 432
        idioma:
          type: string
          example: Español
        libroId:
          type: integer
          format: int64
          example: 1

    # Categorias
...
    LibroCategoriaCreateRequest:
      type: object
      required: [libroId, categoriaId]
      properties:
        libroId:
          type: integer
          format: int64
          example: 1
        categoriaId:
          type: integer
          format: int64
          example: 2
        prioridad:
          type: integer
          example: 1
        comentario:
          type: string
          maxLength: 255
          example: Categoría principal
    LibroCategoriaResponse:
      type: object
      properties:
        libroId:
          type: integer
          format: int64
          example: 1
        libroTitulo:
          type: string
          example: Cien años de soledad
        categoriaId:
          type: integer
          format: int64
          example: 2
        categoriaNombre:
          type: string
          example: Novela
        prioridad:
          type: integer
          example: 1
        addedAt:
          type: string
          format: date-time
        comentario:
          type: string
          example: Categoría principal

    # Roles
...
    UsuarioRolResponse:
      type: object
      properties:
        usuarioId:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: juan123
        rolId:
          type: integer
          format: int64
          example: 2
        rolNombre:
          type: string
          example: ADMIN